Remove external crate type from transport
