+++ dds/src/transport/history_cache.rs
@@ -1,4 +1,4 @@
-use std::sync::Arc;
+use alloc::sync::Arc;
 
 use super::types::{ChangeKind, Guid, Time};
